+++ CHANGELOG.md
@@ -21,7 +21,14 @@ Adding a requirement of a major version of a dependency is breaking a contract.
 Dropping a requirement of a major version of a dependency is a new contract.
 
 ## [Unreleased]
-[Unreleased]: https://github.com/atlassian/aws-infrastructure/compare/release-2.18.0...master
+[Unreleased]: https://github.com/atlassian/aws-infrastructure/compare/release-2.19.0...master
+
+## [2.19.0] - 2019-11-22
+[2.19.0]: https://github.com/atlassian/aws-infrastructure/compare/release-2.18.0...release-2.19.0
+
+### Added
+- Add `instanceType` to `StackVirtualUsersFormula` and `MulticastVirtualUsersFormula`.
+  Allow avoiding low capacity instance types.
 
 ## [2.18.0] - 2019-10-17
 [2.18.0]: https://github.com/atlassian/aws-infrastructure/compare/release-2.17.0...release-2.18.0